All Target files
Target Takes Damage, 
Attacks when in range (however no damage dealt)
LooksAt player (could be changed to rotateTowards for more ~suspense~)

There is a FollowsPlayer script for the Targets, but it doesnt work well, im not sure why. 

Targets move slightly with the player, not stuck on the ground... also dont know why. 

No Spawn script